Replaced `QueuedJobLogger` with `QueuedJobHandler`
